+++ src/lib.rs
@@ -1,4 +1,4 @@
-#![doc(html_root_url = "https://docs.rs/unquote/0.0.5")]
+#![doc(html_root_url = "https://docs.rs/unquote/0.0.6")]
 #![warn(clippy::pedantic)]
 
 #[cfg(doctest)]
@@ -15,6 +15,7 @@ use quote::quote_spanned;
 use std::collections::HashSet;
 use syn::{
 	parse::{ParseStream, Parser},
+	spanned::Spanned,
 	Error, Expr, Ident, Result, Token,
 };
 
@@ -53,12 +54,12 @@ fn unquote_outer(input: ParseStream) -> Result<TokenStream> {
 	let mut declare_up_front = HashSet::new();
 	let output = unquote_inner(input, &input_ident, &mut declare_up_front)?;
 	let declare_up_front = declare_up_front.into_iter();
-	Ok(quote_spanned!(Span::mixed_site()=> {
+	Ok(quote_spanned!(Span::mixed_site()=>
 		let #input_ident = #parse_stream;
 		let mut prev_span = #input_ident.cursor().span();
 		#(let #declare_up_front;)*
 		#output
-	}))
+	))
 }
 
 //TODO: Make errors specific even if the token type is mismatched outright.
@@ -96,9 +97,36 @@ fn unquote_inner(
 					match input.parse().map_err(|_| {
 						Error::new(
 							input.span(),
-							"Unexpected end of macro input: Expected Parse identifier, Span identifier written as lifetime or joined `#`",
+							"Unexpected end of macro input: Expected Parse identifier, Span identifier written as lifetime, joined `#` or `let(pattern)`",
 						)
 					})? {
+						TokenTree::Ident(r#do) if r#do == "do" => {
+							let r#let: Option<Token![let]> = input.parse()?;
+							let parser_function: Expr = input.parse()?;
+							let fat_arrow: Token![=>] = input.parse()?;
+							let placeholder: Ident = input.parse()?;
+							if r#let.is_some() {
+								declare_up_front.insert(placeholder.clone());
+							}
+							hygienic_spanned! {
+								punct.span()
+								.join(r#do.span())
+								.and_then(|s| s.join(fat_arrow.span()))
+								.and_then(|s| s.join(placeholder.span()))
+								.unwrap_or_else(|| r#do.span())
+								=>
+								#placeholder = #input_ident.call(#parser_function)?;
+								prev_span = syn::spanned::Spanned::span(&#placeholder);
+							}
+						}
+						TokenTree::Ident(r#let) if r#let == "let" => {
+							let placeholder: Ident = input.parse()?;
+							declare_up_front.insert(placeholder.clone());
+							hygienic_spanned! {punct.span().join(r#let.span()).and_then(|s| s.join(placeholder.span())).unwrap_or_else(|| r#let.span())=>
+								#placeholder = #input_ident.parse()?;
+								prev_span = syn::spanned::Spanned::span(&#placeholder);
+							}
+						}
 						TokenTree::Ident(placeholder) => {
 							hygienic_spanned! {punct.span().join(placeholder.span()).unwrap_or_else(|| placeholder.span())=>
 								#placeholder = #input_ident.parse()?;